org structure of the codebase
