+++ pipeline/cloud-functions/node/index.js
@@ -1,0 +1,3 @@
+var app = require("./bin/www");
+
+module.exports = { app };
